ui/requirements.txt
streamlit>=1.30.0
azure-monitor-query>=1.3.0
azure-identity>=1.15.0
requests>=2.31.0
pyyaml>=6.0
pandas>=2.0.0
